rename file and creation of an authentication category
