+++ src/public/index.php
@@ -14,9 +14,11 @@ require '../../vendor/autoload.php';
 require '../../src/twig/TwigEnvironment.php';
 require '../controller/PostController.php';
 require '../controller/AdminController.php';
+require '../controller/AuthenticationController.php';
 require '../model/ModelConnect.php';
 require '../model/ModelAdmin.php';
 require '../model/ModelPost.php';
+require '../model/ModelAuthentification.php';
 require '../manager/ManagerValide.php';
 require '../manager/ManagerArticle.php';
 require '../manager/ManagerCommentaire.php';
@@ -25,6 +27,7 @@ require '../manager/ManagerContact.php';
 
 use App\controller\Admincontroller;
 use App\controller\Postcontroller;
+use App\controller\AuthentificationController;
 use Dotenv\Dotenv;
 
 $dotenv = Dotenv::createImmutable('../../');
@@ -33,115 +36,115 @@ $router = new AltoRouter();
 
 $router->map(
     'GET|POST', '/', function () {
-        $page = new postcontroller;
+        $page = new Postcontroller;
         $page->home($_POST);
     }
 );
 $router->map(
     'GET', '/posts', function () {
-        $page = new postcontroller;
+        $page = new Postcontroller;
         $page->allposts();
     }
 );
 $router->map(
     'GET', '/post/[i:id]', function ($id) {
-        $page = new postcontroller;
+        $page = new Postcontroller;
         $page->onepost($id);
     }
 );
 $router->map(
     'GET|POST', '/admin/addpost', function () {
-        $page = new postcontroller;
+        $page = new Postcontroller;
         $page->addUpdate($_POST);
     }
 );
 $router->map(
     'GET|POST', '/updatepost/[i:id]', function ($id) {
-        $page = new postcontroller;
+        $page = new Postcontroller;
         $page->addUpdate($_POST, $id); 
     }
 );
 $router->map(
     'GET', '/removepost/[i:id]', function ($id) {
-        $page = new postcontroller;
+        $page = new Postcontroller;
         $page->remove($id);
     }
 );
 $router->map(
     'GET|POST', '/register', function () {
-        $page = new admincontroller;
+        $page = new AuthentificationController;
         $page->register($_POST);
     }
 );
 $router->map(
     'GET|POST', '/login', function () {
-        $page = new admincontroller;
+        $page = new AuthentificationController;
         $page->login($_POST);
     }
 );
 $router->map(
     'GET', '/logout', function () {
-        $page = new admincontroller;
+        $page = new AuthentificationController;
         $page->logout();
     }
 );
 $router->map(
     'GET|POST', '/admin/roles', function () {
-        $page = new admincontroller;
+        $page = new Admincontroller;
         $page->roles($_POST);
     }
 );
 $router->map(
     'GET', '/admin', function () {
-        $page = new admincontroller;
+        $page = new Admincontroller;
         $page->admin();
     }
 );
 $router->map(
     'POST', '/comment', function () {
-        $page = new postcontroller;
+        $page = new Postcontroller;
         $page->comment($_POST);
     }
 );
 $router->map(
     'GET|POST', '/admin/[a:action]', function ($action) {
-        $page = new admincontroller;
+        $page = new Admincontroller;
         $page->comment($_POST, $action);
     }
 );
 $router->map(
     'GET', '/admin/deletecomment/[i:id]/[a:action]', function ($id,$action) {
-        $page = new admincontroller;
+        $page = new Admincontroller;
         $page->deletecomment($id, $action);
     }
 );
 $router->map(
     'GET', '/admin/deleteuser/[i:id]', function ($id) {
-        $page = new admincontroller;
+        $page = new Admincontroller;
         $page->deleteuser($id);
     }
 );
 $router->map(
     'GET|POST', '/resetpassword', function () {
-        $page = new admincontroller;
+        $page = new AuthentificationController;
         $page->resetpassword($_POST);
     }
 );
 $router->map(
     'GET|POST', '/resetlink/[i:id]/[a:action]', function ($id,$action) {
-        $page = new admincontroller;
+        $page = new AuthentificationController;
         $page->resetlink($id, $action, $_POST);
     }
 );
 $router->map(
     'GET|POST', '/updateuser', function () {
-        $page = new admincontroller;
+        $page = new Admincontroller;
         $page->updateuser($_POST);
     }
 );
 $router->map(
     'GET|POST', '/updatepassword', function () {
-        $page = new admincontroller;
+        $page = new Admincontroller;
         $page->updatepassword($_POST);
     }
 );
